requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.7.4

# Streamlit UI
streamlit==1.28.2
requests==2.31.0

# OpenAI SDK with compatible httpx
openai==1.12.0
httpx==0.27.0

# Vector Store
chromadb==0.4.24

# Sentence Transformers for embeddings
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.35.0

# GitHub API
PyGithub==2.1.1

# Environment Variables
python-dotenv==1.0.0

# Additional Dependencies
numpy==1.24.3